+++ public/js/logic.js
@@ -1,3 +1,7 @@
+
+if (typeof module !== "undefined") {
+    var XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
+}
 function fetchApi(url, cb) {
   var xhr = new XMLHttpRequest();
   xhr.onreadystatechange = function() {
@@ -10,22 +14,6 @@ function fetchApi(url, cb) {
   xhr.send();
 }
 
-//select
-function select(id) {
-    return document.querySelector(id);
-  }
-  //addListener
-  function addListener(selector, eventName, callback) {
-    document.querySelector(selector).addEventListener(eventName, callback);
-  }
-  
-  //createlement
-  
-  function createlement(element) {
-    return document.createElement(element);
-  }
-
-
 if (typeof module !== "undefined") {
-    module.exports = fetchApi;
-  }
+  module.exports = fetchApi;
+}
